fix: migrate isTokenRevoked function to auth-utils.ts
