+++ src/lib/auth-utils.ts
@@ -1,13 +1,18 @@
 import _ from 'lodash';
+import LRU from "lru-cache";
 import { convertPayloadToBase64, ErrorCode } from './utils';
 import { API_ERROR, HTTP_STATUS, ROLES, TIME_EXPIRATION_7D, TOKEN_BASIC, TOKEN_BEARER, DEFAULT_MIN_LIMIT_PASSWORD } from './constants';
 
 import { RemoteUser, Package, Callback, Config, Security, APITokenOptions, JWTOptions, IPluginAuth } from '@verdaccio/types';
 import { CookieSessionToken, IAuthWebUI, IAuth, AuthMiddlewarePayload, AuthTokenHeader, BasicPayload, IStorageHandler } from '../../types';
-import { aesDecrypt, verifyPayload, stringToMD5 } from './crypto-utils';
-import { isTokenRevoked } from "./token-blacklist";
+import { aesDecrypt, verifyPayload, stringToMD5, isUUID } from './crypto-utils';
 
 import { logger } from '../lib/logger';
+
+const blacklistCache = new LRU({
+  max: 10000,
+  maxAge: 1000 * 60 * 60 * 24 // 24h
+});
 
 export function validatePassword(password: string, minLength: number = DEFAULT_MIN_LIMIT_PASSWORD): boolean {
   return typeof password === 'string' && password.length >= minLength;
@@ -17,7 +22,7 @@ export function validatePassword(password: string, minLength: number = DEFAULT_M
  * Create a RemoteUser object
  * @return {Object} { name: xx, pluginGroups: [], real_groups: [] }
  */
-export function createRemoteUser(name: string, pluginGroups: string[]): RemoteUser {
+export function createRemoteUser(name: string | void, pluginGroups: string[]): RemoteUser {
   const isGroupValid: boolean = Array.isArray(pluginGroups);
   const groups = (isGroupValid ? pluginGroups : []).concat([ROLES.$ALL, ROLES.$AUTH, ROLES.DEPRECATED_ALL, ROLES.DEPRECATED_AUTH, ROLES.ALL]);
 
@@ -268,4 +273,35 @@ export async function getMiddlewareCredentials(security: Security, secret: strin
   if (_.isString(token) && scheme.toUpperCase() === TOKEN_BEARER.toUpperCase()) {
     return await verifyJWTPayload(token, secret, storage);
   }
+}
+
+// currently only manually created tokens via the token command can be validated
+// as verdaccio only stores these tokens in the storage handler
+export async function isTokenRevoked(name: string, tokenKey: string, storage: IStorageHandler): Promise<boolean> {
+  if (!isUUID(tokenKey)){
+    // if tokenKey is a UUID it is a unique jwtid and originates from a manually generated jwt token (npm comman)
+    // else it will be an non-identifiable token (jwt or legacy) and can not be validated in absence of an unqiue identifier
+    // we should skip
+    return false;
+  }
+
+  // validate if tokenIdentifier is already listed in memory cache
+  const tokenIdentifier = `${name}:${tokenKey}`;
+  let blacklisted = blacklistCache.get(tokenIdentifier) as boolean | undefined;
+  if (blacklisted !== undefined) { 
+    return blacklisted;
+  }
+
+  // validate if tokenKey is part of user token storage
+  // a fallback lookup; store in memory cache if token is not found
+  try {
+    const tokens = await storage.readTokens({user: name})
+    blacklisted = tokens.find((token) => token.key === tokenKey) === undefined;
+    if (blacklisted) blacklistCache.set(tokenIdentifier, true)
+  } catch (err) {
+    // issue with access to storage pass for now
+    blacklisted = false;
+  }
+
+  return blacklisted;
 }
